ref(api): Move api cleint to own file
